requirements.txt
ImageHash==4.0
numpy==1.16.4
opencv-contrib-python==4.1.0.25
pdf2image==1.6.0
Pillow==6.1.0
pytesseract==0.2.7
PyWavelets==1.0.3
scikit-learn==0.21.3
scipy==1.3.1
requests
torch==1.2.0
torchvision==0.4.0
